agent.py
import os
from typing import Dict, Any, Optional
from dotenv import load_dotenv
from google.adk.agents import Agent
from google.adk.tools.agent_tool import AgentTool
from google.adk.tools import tool

"""
Study Buddy AI Agent

A helpful AI tutor designed to assist students with exam preparation.
Built using Google Agent Development Kit (ADK).
"""

# System Instructions for the Study Buddy AI Agent
ROOT_INSTRUCTIONS = """
You are an experienced and patient tutor who follows a systematic approach to help students master their exam subjects.

## Your Teaching Flow (FOLLOW THIS EXACT ORDER):

### Step 1: Get to Know the Student
Use `get_student_info` to gather:
- Grade level (determines explanation complexity)
- Subject for the exam  
- Exam timeline
- Current knowledge level
- Specific concerns or weak areas

### Step 2: Research the Subject
Use `research_subject` to:
- Understand the curriculum standards for their grade level
- Identify key concepts and topics
- Determine typical difficulty areas
- Gather relevant teaching materials and examples

### Step 3: Divide Subject into Manageable Sections
Use `create_study_sections` to:
- Break the subject into logical, sequential sections
- Ensure each section builds on the previous one
- Set clear learning objectives for each section
- Create a realistic timeline based on exam date

### Step 4: For Each Section - Teaching Loop:

#### 4a: Teach the Material
Use `teach_section` to:
- Explain concepts clearly with age-appropriate language
- Provide real-world examples and analogies
- Start with fundamentals, build complexity gradually
- Use interactive explanations suited to their learning style

#### 4b: Quiz the Student  
Use `quiz_section` to:
- Test understanding with targeted questions
- Include various question types (multiple choice, short answer, problem-solving)
- Assess both recall and application of concepts
- Make questions appropriate for their grade level

#### 4c: Check Understanding
Use `check_understanding` to:
- Analyze quiz performance
- Identify specific concepts that need more work
- Determine if student is ready for next section

#### 4d: Decision Point:
- **IF understanding is sufficient**: Move to next section (repeat 4a-4d)
- **IF understanding is lacking**: Go through weak concepts again, then re-quiz (repeat 4a-4c)

### Key Teaching Principles:
- Be patient and encouraging - learning takes time
- Adapt explanations to student's grade level and learning style  
- Never move forward until current section is mastered
- Celebrate progress and build confidence
- Use positive reinforcement throughout

### Communication Style:
- Use encouraging, supportive language
- Ask questions to check understanding frequently
- Provide clear, step-by-step explanations
- Give specific, actionable feedback
- Make learning engaging and relevant

Your goal is complete mastery of the subject, not just passing the exam. Take as much time as needed in each section to ensure solid understanding before moving forward.
"""


# Load environment variables
load_dotenv()


# Tools for the Study Plan Agent Flow

def get_student_info(grade_level: str, subject: str, exam_date: str, current_knowledge: str) -> dict:
    """
    Collect essential information about the student and their learning needs.
    
    Args:
        grade_level: Student's current grade (e.g., "9th grade", "college freshman")
        subject: Subject for the exam (e.g., "Biology", "Algebra II")
        exam_date: When the exam is scheduled
        current_knowledge: Student's self-assessed knowledge level
        
    Returns:
        dict: Structured student profile
    """
    return {
        "student_profile": {
            "grade_level": grade_level,
            "subject": subject,
            "exam_date": exam_date,
            "current_knowledge": current_knowledge,
            "profile_complete": True
        },
        "next_step": "research_subject"
    }


def research_subject(subject: str, grade_level: str) -> dict:
    """
    Research the subject to understand key concepts and curriculum standards.
    
    Args:
        subject: The subject to research
        grade_level: Student's grade level to determine appropriate depth
        
    Returns:
        dict: Subject research summary and key concepts
    """
    # This would integrate with knowledge bases, curriculum standards, etc.
    return {
        "subject_overview": f"Research completed for {subject} at {grade_level} level",
        "key_concepts": [
            "Fundamental principles",
            "Core theories and applications", 
            "Problem-solving techniques",
            "Common exam topics"
        ],
        "difficulty_areas": ["Abstract concepts", "Mathematical applications", "Memorization requirements"],
        "next_step": "create_study_sections"
    }


def create_study_sections(subject: str, key_concepts: list, exam_date: str) -> dict:
    """
    Divide the subject into manageable study sections with logical progression.
    
    Args:
        subject: The subject being studied
        key_concepts: List of key concepts from research
        exam_date: Target exam date for scheduling
        
    Returns:
        dict: Organized study sections with timeline
    """
    sections = []
    for i, concept in enumerate(key_concepts, 1):
        sections.append({
            "section_number": i,
            "title": concept,
            "estimated_time": "3-5 study sessions",
            "prerequisites": f"Understanding of previous {i-1} sections" if i > 1 else "Basic foundational knowledge",
            "learning_objectives": [f"Master {concept}", f"Apply {concept} to problems", f"Explain {concept} clearly"],
            "status": "not_started"
        })
    
    return {
        "study_sections": sections,
        "total_sections": len(sections),
        "current_section": 1,
        "next_step": "teach_section"
    }


def teach_section(section_number: int, section_title: str, student_grade: str) -> dict:
    """
    Teach the material for a specific section with explanations and examples.
    
    Args:
        section_number: Which section we're teaching
        section_title: Title/topic of the section
        student_grade: Student's grade level for appropriate explanations
        
    Returns:
        dict: Teaching content and next steps
    """
    return {
        "section_taught": {
            "number": section_number,
            "title": section_title,
            "explanations_provided": True,
            "examples_given": True,
            "grade_appropriate": student_grade
        },
        "teaching_summary": f"Completed teaching {section_title} with explanations and examples suitable for {student_grade}",
        "next_step": "quiz_section"
    }


def quiz_section(section_number: int, section_title: str, difficulty_level: str = "medium") -> dict:
    """
    Create and administer a quiz to test understanding of the section.
    
    Args:
        section_number: Which section to quiz on
        section_title: Title of the section being quizzed
        difficulty_level: How challenging the quiz should be
        
    Returns:
        dict: Quiz results and comprehension assessment
    """
    return {
        "quiz_administered": {
            "section": section_number,
            "title": section_title,
            "difficulty": difficulty_level,
            "questions_asked": 5,
            "answers_received": True
        },
        "next_step": "check_understanding"
    }


def check_understanding(quiz_results: dict, passing_threshold: float = 0.8) -> dict:
    """
    Analyze quiz results to determine if student understands the material sufficiently.
    
    Args:
        quiz_results: Results from the section quiz
        passing_threshold: Minimum score to consider understanding adequate (0.0-1.0)
        
    Returns:
        dict: Understanding assessment and next action
    """
    # This would analyze actual quiz performance
    # For now, we'll simulate the decision-making process
    
    understanding_level = 0.75  # This would be calculated from actual quiz results
    
    if understanding_level >= passing_threshold:
        return {
            "understanding_sufficient": True,
            "score": understanding_level,
            "next_action": "proceed_to_next_section",
            "message": "Great job! You've mastered this section. Let's move to the next one."
        }
    else:
        return {
            "understanding_sufficient": False,
            "score": understanding_level,
            "next_action": "review_and_reteach",
            "weak_concepts": ["Concept A", "Concept B"],  # Would be identified from quiz analysis
            "message": "Let's review some concepts you're still working on, then try the quiz again."
        }


root_agent = Agent(
    name='study_buddy_tutor',
    model='gemini-2.0-flash-exp',
    instruction=ROOT_INSTRUCTIONS,
    tools=[
        get_student_info,
        research_subject, 
        create_study_sections,
        teach_section,
        quiz_section,
        check_understanding
    ]
)
__all__ = ['root_agent']