pyproject.toml
[project]
name = "study-buddy-ai"
version = "0.1.0"
description = "A study buddy AI agent for students using Google ADK"
authors = [
    {name = "Elijah", email = "your-email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "flake8",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "black",
    "flake8",
    "isort",
    "mypy",
    "pytest",
]
